add support for async functions.
A Function is a block of code which can return an error if it is not
successful. To call this function, one can use MustSucceed which tells
nsq to invoke the handler until the function itself returns no error.

The publisher and consumer are both implemented by the using process so
this function can be used to async call a function. If your process is
running inside a cluster it is possible to distribute the invocation to
other instances.